+++ src/utils/itr.js
@@ -1,0 +1,6 @@
+import toIterable from 'browser-readablestream-to-it'
+
+export function asAsyncIterable (readable) {
+  return Symbol.asyncIterator in readable ? readable : toIterable(readable)
+}
+
